Add C++ style message deleter
Allows to use any callable that can be called as `void deleter(void *buf)` as zeromq message deleter.
In case plain static function (or lambda convertible to a such function) is passed as deleter no extra allocation is performed.
It is also possible to use ZeroMQ approach and pass function (or lambda convertible to a such function) with signature
`void deleter(void *buf, void *hint)` and a hint (aka user data).